backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
alembic==1.13.1
pillow==10.1.0
python-multipart==0.0.6
tenacity==8.2.3
apscheduler==3.10.4
sqlalchemy-utils==0.41.1
psutil==5.9.5
psycopg2-binary==2.9.9
# GOÄ-Rechnungsmodul
weasyprint==62.3
factur-x==3.7
qrcode[pil]==8.0
jinja2==3.1.4
